Split out State from MobiusController (and restore comment)
