+++ MobiusCore/Source/MobiusController.swift
@@ -28,6 +28,20 @@ public final class MobiusController<Model, Event, Effect> {
     typealias ViewConnectable = AsyncDispatchQueueConnectable<Model, Event>
     typealias ViewConnection = Connection<Model>
 
+    private struct StoppedState {
+        var modelToStartFrom: Model
+        var viewConnectable: ViewConnectable?
+
+    }
+
+    private struct RunningState {
+        var loop: Loop
+        var viewConnectable: ViewConnectable?
+        var disposables: CompositeDisposable
+    }
+
+    private typealias State = AsyncStartStopStateMachine<StoppedState, RunningState>
+
     private let loopFactory: (Model) -> Loop
     private let loopQueue: DispatchQueue
     private let viewQueue: DispatchQueue
@@ -47,6 +61,30 @@ public final class MobiusController<Model, Event, Effect> {
         loopQueue loopTargetQueue: DispatchQueue,
         viewQueue: DispatchQueue
     ) {
+        /*
+         Ownership graph after initing:
+
+                     ┏━━━━━━━━━━━━┓
+                ┌────┨ controller ┠────────┬──┐
+                │    ┗━━━━━━━━━━┯━┛        │  │
+         ┏━━━━━━┷━━━━━━┓    ┏━━━┷━━━━━━━┓  │  │
+         ┃ loopFactory ┃    ┃ viewQueue ┃  │  │
+         ┗━━━━━━━━━┯━━━┛    ┗━━━━━━━━━━━┛  │  │
+              ┏━━━━┷━━━━━━━━━━━━━━━━━━┓    │  │
+              ┃ flipEventsToLoopQueue ┃ ┌──┘  │
+              ┗━━━━━━━━━━━━━━┯━━━━━━┯━┛ │     │
+                             │    ┏━┷━━━┷━┓   │
+                             │    ┃ state ┃   │
+                             │    ┗━┯━━━━━┛   │
+                             │      │ ┌───────┘
+                           ┏━┷━━━━━━┷━┷┓
+                           ┃ loopQueue ┃
+                           ┗━━━━━━━━━━━┛
+
+         In order to construct this bottom-up and fulfil definitive initialization requirements, state and loopQueue are
+         duplicated in local variables.
+         */
+
         // The internal loopQueue is a serial queue targeting the provided queue, so that targeting a concurrent queue
         // doesn’t result in concurrent work on the underlying MobiusLoop. This behaviour is documented on
         // `Mobius.Builder.makeController`.
@@ -54,7 +92,10 @@ public final class MobiusController<Model, Event, Effect> {
         self.loopQueue = loopQueue
         self.viewQueue = viewQueue
 
-        let state = State(model: initialModel, queue: loopQueue)
+        let state = State(
+            state: StoppedState(modelToStartFrom: initialModel, viewConnectable: nil),
+            queue: loopQueue
+        )
         self.state = state
 
         // Maps an event consumer to a new event consumer that invokes the original one on the loop queue,
@@ -213,165 +254,6 @@ public final class MobiusController<Model, Event, Effect> {
             case .running(let state):
                 return state.loop.latestModel
             }
-        }
-    }
-
-    // MARK: - State model
-
-    private struct StoppedState {
-        var modelToStartFrom: Model
-        var viewConnectable: ViewConnectable?
-
-    }
-
-    private struct RunningState {
-        var loop: Loop
-        var viewConnectable: ViewConnectable?
-        var disposables: CompositeDisposable
-    }
-
-    // State machine representing the two states of a controller, stopped and running.
-    // This isn’t an enum because we need synchronized access to the running flag without copying around arbitrarily
-    // large models.
-    //
-    // `running` can safely be read from any thread. Changing `running` and reading or writing state is restricted
-    // to the loop queue.
-    private final class State {
-        // Intermediate states are required to handle the fun, fun case where an event source dispatches an event
-        // immediately, which happens before the loop variable is assigned in start(). In this case, we enter
-        // flipEventsToLoopQueue() in the .transitioningToRunning state, but the async block cannot proceed until we
-        // reach the .running state (because start() is already running on the loop queue).
-        private enum RawState {
-            case stopped
-            case transitioningToRunning
-            case running
-            case transitioningToStopped
-        }
-
-        private let rawState = Synchronized(value: RawState.stopped)
-        private let loopQueue: DispatchQueue
-        private var stoppedState: StoppedState?
-        private var runningState: RunningState?
-
-        enum Snapshot {
-            case stopped(StoppedState)
-            case running(RunningState)
-        }
-
-        init(model: Model, queue: DispatchQueue) {
-            stoppedState = StoppedState(modelToStartFrom: model, viewConnectable: nil)
-            loopQueue = queue
-        }
-
-        /// Test whether we’re in a running state. Ongoing transitions are considered running states.
-        ///
-        /// This is safe to invoke from any queue, including the loop queue.
-        var running: Bool {
-            switch rawState.value {
-            case .stopped:
-                return false
-            case .transitioningToRunning, .running, .transitioningToStopped:
-                return true
-            }
-        }
-
-        /// Call `closure` with the current state. It will execute on the loop queue.
-        func syncRead<T>(_ closure: (Snapshot) throws -> T) rethrows -> T {
-            dispatchPrecondition(condition: .notOnQueue(loopQueue))
-            return try loopQueue.sync {
-                try closure(snapshot())
-            }
-        }
-
-        /// Mutate the stopped state, assuming we’re currently stopped. If not, fail with the provided error message.
-        func mutateIfStopped(elseError error: String, _ closure: (inout StoppedState) -> Void) {
-            dispatchPrecondition(condition: .notOnQueue(loopQueue))
-            loopQueue.sync {
-                switch snapshot() {
-                case .running:
-                    MobiusHooks.onError(error)
-                case .stopped(var state):
-                    closure(&state)
-                    stoppedState = state
-                }
-            }
-        }
-
-        /// Transition from a stopped state to a running state, assuming we’re currently stopped. If not, fail with the
-        /// provided error message.
-        ///
-        /// The `transition` closure may return nil to indicate failure, in which case the state remains unchanged. This
-        /// behaviour isn’t desired but is forced by our error hooking mechanism – if `transition` calls
-        /// `MobiusHooks.onError` it should then return `nil`.
-        func transitionToRunning(elseError error: String, _ transition: (StoppedState) -> RunningState?) {
-            dispatchPrecondition(condition: .notOnQueue(loopQueue))
-            loopQueue.sync {
-                switch snapshot() {
-                case .running:
-                    MobiusHooks.onError(error)
-                case .stopped(let stoppedState):
-                    rawState.value = .transitioningToRunning
-                    if let runningState = transition(stoppedState) {
-                        become(running: runningState)
-                    } else {
-                        rawState.value = .stopped
-                    }
-                }
-            }
-        }
-
-        /// Transition from a running state to a stopped state, assuming we’re currently running. If not, fail with the
-        /// provided error message.
-        ///
-        /// The `transition` closure may return nil to indicate failure, in which case the state remains unchanged. This
-        /// behaviour isn’t desired but is forced by our error hooking mechanism – if `transition` calls
-        /// `MobiusHooks.onError` it should then return `nil`.
-        func transitionToStopped(elseError error: String, _ transition: (RunningState) -> StoppedState?) {
-            dispatchPrecondition(condition: .notOnQueue(loopQueue))
-            loopQueue.sync {
-                switch snapshot() {
-                case .stopped:
-                    MobiusHooks.onError(error)
-                case .running(let runningState):
-                    rawState.value = .transitioningToStopped
-                    if let stoppedState = transition(runningState) {
-                        become(stopped: stoppedState)
-                    } else {
-                        rawState.value = .running
-                    }
-                }
-            }
-        }
-
-        /// Generate a `Snapshot`  reflecting the current, er, state of the `State`.
-        ///
-        /// This function is the only point where we deal with the two optionals.
-        private func snapshot() -> Snapshot {
-            dispatchPrecondition(condition: .onQueue(loopQueue))
-
-            if running {
-                guard let runningState = runningState else { preconditionFailure("Internal invariant broken") }
-                return .running(runningState)
-            } else {
-                guard let stoppedState = stoppedState else { preconditionFailure("Internal invariant broken") }
-                return .stopped(stoppedState)
-            }
-        }
-
-        private func become(running state: RunningState) {
-            dispatchPrecondition(condition: .onQueue(loopQueue))
-
-            self.runningState = state
-            rawState.value = .running
-            self.stoppedState = nil
-        }
-
-        private func become(stopped state: StoppedState) {
-            dispatchPrecondition(condition: .onQueue(loopQueue))
-
-            self.stoppedState = state
-            rawState.value = .stopped
-            self.runningState = nil
         }
     }
 }
